+++ packages/qti-components/CHANGELOG.md
@@ -1,5 +1,98 @@
 # Changelog
 
+## [7.20.4](https://github.com/Citolab/qti-components/compare/qti-components-v7.20.3...qti-components-v7.20.4) (2026-02-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @qti-components/interactions bumped to 1.4.2
+    * @qti-components/item bumped to 1.0.6
+    * @qti-components/test bumped to 1.1.5
+
+## [7.20.3](https://github.com/Citolab/qti-components/compare/qti-components-v7.20.2...qti-components-v7.20.3) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @qti-components/interactions bumped to 1.4.1
+    * @qti-components/item bumped to 1.0.5
+    * @qti-components/test bumped to 1.1.4
+
+## [7.20.2](https://github.com/Citolab/qti-components/compare/qti-components-v7.20.1...qti-components-v7.20.2) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @qti-components/interactions bumped to 1.4.0
+    * @qti-components/item bumped to 1.0.4
+    * @qti-components/test bumped to 1.1.3
+
+## [7.20.1](https://github.com/Citolab/qti-components/compare/qti-components-v7.20.0...qti-components-v7.20.1) (2026-02-04)
+
+
+### Bug Fixes
+
+* add private method for positioning custom menu ([0a596de](https://github.com/Citolab/qti-components/commit/0a596de0db4c23be54b4c90592918cf1f5c8c929))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @qti-components/interactions bumped to 1.3.1
+    * @qti-components/item bumped to 1.0.3
+    * @qti-components/test bumped to 1.1.2
+
+## [7.20.0](https://github.com/Citolab/qti-components/compare/qti-components-v7.19.0...qti-components-v7.20.0) (2026-02-03)
+
+
+### Features
+
+* add private methods for response handling and normalization ([a763136](https://github.com/Citolab/qti-components/commit/a7631361a728df10dfd3cbb37114075505ac7b00))
+* add private methods for response handling and normalization ([9a5e165](https://github.com/Citolab/qti-components/commit/9a5e165600c67276cd0108f2de9c581c1e80b404))
+
+## [7.19.0](https://github.com/Citolab/qti-components/compare/qti-components-v7.18.1...qti-components-v7.19.0) (2026-02-03)
+
+
+### Features
+
+* enhance qti-inline-choice-interaction ([5878343](https://github.com/Citolab/qti-components/commit/58783433b0f47dcf31bc75de0b6427c98860c2df))
+
+
+### Bug Fixes
+
+* inline choice dropdown font size and positioning on mobile devices ([52d3216](https://github.com/Citolab/qti-components/commit/52d3216a0661f863b0f8384763bd829d6a41a19a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @qti-components/interactions bumped to 1.3.0
+    * @qti-components/item bumped to 1.0.2
+    * @qti-components/test bumped to 1.1.1
+    * @qti-components/theme bumped to 1.0.1
+
+## [7.18.1](https://github.com/Citolab/qti-components/compare/qti-components-v7.18.0...qti-components-v7.18.1) (2026-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * devDependencies
+    * @qti-components/elements bumped to 1.0.1
+    * @qti-components/interactions bumped to 1.2.0
+    * @qti-components/item bumped to 1.0.1
+    * @qti-components/loader bumped to 1.0.1
+    * @qti-components/test bumped to 1.1.0
+    * @qti-components/transformers bumped to 1.2.0
+
 ## [7.18.0](https://github.com/Citolab/qti-components/compare/qti-components-v7.17.1...qti-components-v7.18.0) (2026-02-02)