+++ packages/qti-item/CHANGELOG.md
@@ -1,5 +1,68 @@
 # Changelog
 
+## [1.0.6](https://github.com/Citolab/qti-components/compare/item-v1.0.5...item-v1.0.6) (2026-02-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.4.2
+
+## [1.0.5](https://github.com/Citolab/qti-components/compare/item-v1.0.4...item-v1.0.5) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.4.1
+
+## [1.0.4](https://github.com/Citolab/qti-components/compare/item-v1.0.3...item-v1.0.4) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.4.0
+
+## [1.0.3](https://github.com/Citolab/qti-components/compare/item-v1.0.2...item-v1.0.3) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.3.1
+
+## [1.0.2](https://github.com/Citolab/qti-components/compare/item-v1.0.1...item-v1.0.2) (2026-02-03)
+
+
+### Bug Fixes
+
+* fix tests ([7a75aa9](https://github.com/Citolab/qti-components/commit/7a75aa9222be3e3b68dbe18ef8895a968903d10c))
+* inline choice dropdown font size and positioning on mobile devices ([52d3216](https://github.com/Citolab/qti-components/commit/52d3216a0661f863b0f8384763bd829d6a41a19a))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.3.0
+    * @qti-components/theme bumped to 1.0.1
+
+## [1.0.1](https://github.com/Citolab/qti-components/compare/item-v1.0.1...item-v1.0.1) (2026-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/elements bumped to 1.0.1
+    * @qti-components/interactions bumped to 1.2.0
+    * @qti-components/transformers bumped to 1.2.0
+
 ## [1.0.1](https://github.com/Citolab/qti-components/compare/item-v1.0.0...item-v1.0.1) (2026-02-02)