+++ packages/qti-test/CHANGELOG.md
@@ -1,5 +1,68 @@
 # Changelog
 
+## [1.1.5](https://github.com/Citolab/qti-components/compare/test-v1.1.4...test-v1.1.5) (2026-02-05)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.4.2
+
+## [1.1.4](https://github.com/Citolab/qti-components/compare/test-v1.1.3...test-v1.1.4) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.4.1
+
+## [1.1.3](https://github.com/Citolab/qti-components/compare/test-v1.1.2...test-v1.1.3) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.4.0
+
+## [1.1.2](https://github.com/Citolab/qti-components/compare/test-v1.1.1...test-v1.1.2) (2026-02-04)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.3.1
+
+## [1.1.1](https://github.com/Citolab/qti-components/compare/test-v1.1.0...test-v1.1.1) (2026-02-03)
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/interactions bumped to 1.3.0
+    * @qti-components/theme bumped to 1.0.1
+
+## [1.1.0](https://github.com/Citolab/qti-components/compare/test-v1.0.1...test-v1.1.0) (2026-02-03)
+
+
+### Features
+
+* enhance PCI state handling with serialization and restoration s… ([f56bc31](https://github.com/Citolab/qti-components/commit/f56bc311c24ef200ab49ebdcb2a0e5771c200f1a))
+* enhance PCI state handling with serialization and restoration support ([5a4638c](https://github.com/Citolab/qti-components/commit/5a4638c9bf6c0b246aba8f5d719852cad73b9757))
+
+
+### Dependencies
+
+* The following workspace dependencies were updated
+  * dependencies
+    * @qti-components/elements bumped to 1.0.1
+    * @qti-components/interactions bumped to 1.2.0
+    * @qti-components/transformers bumped to 1.2.0
+
 ## [1.0.1](https://github.com/Citolab/qti-components/compare/test-v1.0.0...test-v1.0.1) (2026-02-02)